+++ integrations/resend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @trigger.dev/resend
 
+## 2.0.6
+
+### Patch Changes
+
+- @trigger.dev/integration-kit@2.0.6
+- @trigger.dev/sdk@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- @trigger.dev/integration-kit@2.0.5
+- @trigger.dev/sdk@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes