+++ integrations/supabase/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @trigger.dev/supabase
 
+## 2.0.6
+
+### Patch Changes
+
+- @trigger.dev/integration-kit@2.0.6
+- @trigger.dev/sdk@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- 3cf3eaff: You can now trigger on multiple database events in the same job
+  - @trigger.dev/integration-kit@2.0.5
+  - @trigger.dev/sdk@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes