+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-trigger
 
+## 2.0.6
+
+## 2.0.5
+
 ## 2.0.4
 
 ### Patch Changes