+++ packages/trigger-sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @trigger.dev/sdk
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [59075f5f]
+  - @trigger.dev/core@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- @trigger.dev/core@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes